Avoid duplicate counts from search
Existing search with keyword will count one entry multiple times if the search keyword appears in multiple field for same entry.

For example, there are in total 1247 entries in `colors.json`

```sh
$ grep -c hsvValue servers/flask/data/colors.json
1247
```

But search with keyword '%' would return 8729 in total which is more than all entries in the json.

```sh
$ curl -s "http://localhost:44555/api/colors?page=1&search=%&size=30&timestamp=2017-07-06T01%3A12%3A55.662Z" | jq .total
8729
```